# doc/api.yml
openapi: "3.0.1"
info:
  title: "OpenAPI definition"
  version: "v0"

servers:
  - url: "https://lfwt56dciff7biiy2ibilezrnm0hanrh.lambda-url.eu-west-3.on.aws"
    description: "Generated server url"

paths:
  /payments/search:
    put:
      tags:
        - payment-controller
      operationId: getPayments
      parameters:
        - name: apiKey
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/PaymentInfo"
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"

  /orange/sync:
    put:
      tags:
        - payment-controller
      operationId: sync
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/RecoveryResult"

  /payment:
    get:
      tags:
        - payment-controller
      operationId: getPayment
      parameters:
        - name: apiKey
          in: query
          required: true
          schema:
            type: string
        - name: payerEmail
          in: query
          required: true
          schema:
            type: string
        - name: pspType
          in: query
          required: true
          schema:
            type: string
            enum:
              - ORANGE_MONEY
        - name: pspPaymentId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Payment"

    post:
      tags:
        - payment-controller
      operationId: createPayment
      parameters:
        - name: apiKey
          in: query
          required: true
          schema:
            type: string
        - name: payerEmail
          in: query
          required: true
          schema:
            type: string
        - name: pspType
          in: query
          required: true
          schema:
            type: string
            enum:
              - ORANGE_MONEY
        - name: pspPaymentId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Payment"

  /orange/transactions/import:
    post:
      tags:
        - payment-controller
      operationId: saveTransactions
      parameters:
        - name: apiKey
          in: query
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - excel
              properties:
                excel:
                  type: string
                  format: binary
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/ImportedTransactionDetails"

  /ping:
    get:
      tags:
        - ping-controller
      operationId: ping
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: string

  /health/email:
    get:
      tags:
        - health-email-controller
      operationId: send_emails
      parameters:
        - name: to
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: string

  /health/bucket:
    get:
      tags:
        - health-bucket-controller
      operationId: file_can_be_uploaded_then_signed
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: string

components:
  schemas:
    PaymentInfo:
      type: object
      properties:
        payerEmail:
          type: string
        pspType:
          type: string
          enum:
            - ORANGE_MONEY
        pspPaymentId:
          type: string

    Application:
      type: object
      properties:
        name:
          type: string
        apiKey:
          type: string

    Payment:
      type: object
      properties:
        id:
          type: string
        pspPayment:
          $ref: "#/components/schemas/PspPayment"
        creationInstant:
          type: string
          format: date-time
        lastPspVerificationInstant:
          type: string
          format: date-time
        verificationAttemptNb:
          type: integer
          format: int32
        payer:
          $ref: "#/components/schemas/User"
        application:
          $ref: "#/components/schemas/Application"
        verificationStatus:
          type: string
          enum:
            - VERIFYING
            - SUCCEEDED
            - FAILED

    PspPayment:
      type: object
      properties:
        pspType:
          type: string
          enum:
            - ORANGE_MONEY
        id:
          type: string
        amount:
          type: integer
          format: int32
        creationInstant:
          type: string
          format: date-time

    User:
      type: object
      properties:
        email:
          type: string

    RecoveryResult:
      type: object
      properties:
        date:
          type: string
          format: date
        isSuccessful:
          type: boolean
        nbSyncedTransactions:
          type: integer
          format: int32
        nbTotalTransactions:
          type: integer
          format: int32
        errorMessage:
          type: string

    ImportedTransactionDetails:
      type: object
      properties:
        bucketKey:
          type: string
        importDate:
          type: string
          format: date-time
        fileName:
          type: string